A12b: convert flow edges once, and stop letting a move edit its own definition
Two structural repairs on the authored-technique runtime representation, plus
the edge enumeration that claimed to be one authority and was three.

FLOW EDGES ARE THE CURSOR'S OWN u16. `FlowNode::{Emit,Wait,Branch}` wrote their
targets as `usize` and the interpreter narrowed with `*then as u16` on every
transition. `MovePlayback::flow_node` is a `u16`, so an authored edge of 65,536
became node 0 at runtime: a terminating flow turned into a loop, its `Emit`
firing again every tick for the rest of the move, with every edge still in range
and nothing for the dangling-edge check to report. A12a's 256-node bound closed
this by sitting under the cliff — which made one number do two jobs. The width
does the second one structurally now, so the node limit is a budget again and a
version 2 may raise it on its own merits. Every in-repo flow is built from Rust
literals, where the value was already a compile error; the road the defect was
actually reachable on is the one this contract exists for — a character LOADED
FROM DATA — so the witness is at the deserialization boundary. Poison-verified
by putting `Emit`'s edge back to `usize`: the test fails with the 65,536 value
accepted.

THE PLAYED DEFINITION IS SHARED, NOT OWNED. `MovePlayback::spec` is an
`Arc<MoveSpec>`. The admission contract's rule is that existing moves keep
referencing the exact prepared definition they started with; while the playback
owned a `MoveSpec` nothing enforced it, because `advance_move_playback` holds
`&mut MovePlayback` every tick — a move's own interpreter could edit the
definition it was executing. `Arc` has no `DerefMut`, so that is now E0594
rather than a rule somebody has to keep, and every one of the read sites was
carried unchanged by `Deref`. It also deletes the per-tick graph clone: the flow
loop deep-copied the whole authored graph, every node and every `EffectRef` key
string, on each tick of each move that authors one — paid again for every
resimulated rollback frame — purely to split the borrow that writing `flow_node`
needs. A refcount bump buys the same split.

ONE EDGE ENUMERATION, AS ITS OWN DOC ALREADY CLAIMED. `TechniqueFlow::successors`
says "ONE PLACE THE EDGES ARE ENUMERATED ... three copies of a match over
`FlowNode` is how a fourth variant comes to be checked by two of them" — while
`reaches_finish` walked its own arms and `problems`' dangling report walked a
third, in the same impl. Both go through `successors` now, which carries the
authored field name along so the diagnostic still names `then` / `on_timeout` /
`otherwise`.

Lane: scripts/run_tests.py --rust, 6/6 jobs, exit 0. Not run: the feature-gated
union, the external-consumer fixtures, the wasm link.

+++ crates/ambition_combat/src/moveset/tests.rs
@@ -8934,7 +8934,7 @@ mod technique_flow {
         }
     }
 
-    fn effect(key: &str, then: usize) -> FlowNode {
+    fn effect(key: &str, then: u16) -> FlowNode {
         FlowNode::Emit {
             effect: ambition_entity_catalog::EffectRef {
                 key: key.to_string(),
